Implement a new WebSocket handler
